+++ capybara/.rspec
@@ -1,2 +1,7 @@
 --color
 --require spec_helper
+--format html
+--out rspec_results.html
+--format j
+--out rspec_results.json
+--format documentation
