+++ capybara/spec/spec_helper.rb
@@ -6,7 +6,8 @@ require 'capybara'
 require 'capybara/rspec'
 require 'site_prism'
 
-require_relative 'support/pages/support_classes.rb'
+#require_relative 'support/pages/support_classes.rb'
+require_relative 'support/pages/all_page_objects.rb'
 
 Capybara.default_driver = :selenium
 Capybara.default_max_wait_time = 20
